requirements.txt
scapy
python-nmap
graphviz
pycryptodome
requests
pyshark
pillow
ttkbootstrap 